feat(anchor): add anchor to tor nav icon
